refactor(test): move findFermentArtifact to shared assertions helper
Deduplicate the polling helper that was copied verbatim into both
ferment-plan-review-oneshot.test.ts and ferment-plan-review-suppression.test.ts.
Export it from tests/e2e/tui/support/assertions.ts and import it in both callers.
